Add the media_profile.xml
Lenovo (or MTK?) defined the media profile differently. Since we don't
have their source, we do it in the standard way. This is essentil.
Without the media profile, standard Camera and media services don't have
to initial values for codecs.
This is derived from Nexus devices. Not fully working yet. Need to
review later.